+++ src/main/java/online/kingdomkeys/kingdomkeys/network/cts/CSUseReactionCommandPacket.java
@@ -18,6 +18,7 @@ import online.kingdomkeys.kingdomkeys.network.PacketHandler;
 import online.kingdomkeys.kingdomkeys.network.stc.SCSyncPlayerData;
 import online.kingdomkeys.kingdomkeys.reactioncommands.ModReactionCommands;
 import online.kingdomkeys.kingdomkeys.reactioncommands.ReactionCommand;
+import online.kingdomkeys.kingdomkeys.util.Utils;
 
 public record CSUseReactionCommandPacket(int index, int lockedOnEntity) implements Packet {
 
@@ -38,8 +39,7 @@ public record CSUseReactionCommandPacket(int index, int lockedOnEntity) implemen
 	@Override
 	public void handle(IPayloadContext context) {
 		Player player = context.player();
-		PlayerData playerData = PlayerData.get(player);
-		String reactionName = playerData.getReactionCommands().get(index);
+		String reactionName = Utils.getRCNameFromIndex(player, index);
 		ReactionCommand reaction = ModReactionCommands.registry.get(ResourceLocation.parse(reactionName));
         if (NeoForge.EVENT_BUS.post(new ReactionCommandCastEvent(player, ResourceLocation.parse(reactionName))).isCanceled())
             return;