+++ src/main/java/online/kingdomkeys/kingdomkeys/reactioncommands/ReactionMagic.java
@@ -14,7 +14,7 @@ public class ReactionMagic extends ReactionCommand {
 	ResourceLocation magic;
 
 	public ReactionMagic(ResourceLocation registryName) {
-		super(registryName, false);
+		super(registryName, false, 30*20);
 		this.magic = registryName;		
 	}