+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 **Added:**
 * New data version check, will automatically run the exp fix command if the playerdata is older.
 * Keyblade swapping by cancelling the ATTACK menu (opposite of opening the submenu).
+* Situation commands (timed reaction commands).
 * Some abilities to some elemental Keyblades.
 * New ability Superjump, increases vertical impulse off a flowmotion jump.
   * Obtained at Valor form levels 4 and 6.